// backend/src/modules/auth/auth.schema.ts
import { z } from 'zod'

export const signupSchema = z.object({
  name: z.string().min(2).max(60),
  email: z.string().email(),
  password: z.string().min(8).max(128)
})

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string()
})

export type SignupInput = z.infer<typeof signupSchema>
export type LoginInput = z.infer<typeof loginSchema>